Terrain Generation Debugging Diagnostic
In a random terrain, two blocks were found colliding.  I included a
comment block that can be reenabled with a single forward slash to test
axis aligned rendering, so that the issue is known to be in the
generation of the terrain itself, not in my rendering code.  I'll look
at the generation code tomorrow, but it's after midnight right now.