Potential TerraGen Fix
Due to improper checking bounds a hypothetical bit of terrain should
have been able to collide with its previous index.  This may have been
the cause of the issue, but the TerraGen bug will be considered to still
exist for the next 48 hours, at which time it will be closed if no test
case is found in which two bits of terrain collide.